refactor: module stuff to the correct place
