--- app/native/src/api/download/mod.rs
@@ -1,4 +1,0 @@
-pub mod assets;
-pub mod library;
-pub mod rules;
-pub mod util;
